Add pipeline splitting logic for Curitiba
By a request, adding a logic to split the Curitiba pipeline (1077) in two: 1077 and 2184, one for TV and the other for Radio and TOPVIEW.
This was made hardcode because creating a new pipeline in AdSim is not possible.
The new pipeline Curitiba 2 was also created hardcode in database, the new logic in script maps each user in Curitiba and assign the correct pipeline id. For every new user a new ID should be inserted in dictionary.
I used the responsible id for the deal to map the pipeline.